fix(cloudtrail-ingest-pipeline): process resources correctly
The Cloudtrail event field `resources` contains an array. So it can
contain multiple objects. Until now the Cloudtrail Ingest Pipeline
did execute a `Rename` processor for every field under `resources`.

This actually does not work because it is an array. This is fixed with
this commit. Now the first object in this array is parsed into the
resources field. So the current structure of the Elasticsearch document
will not change.

To indicate events which are associated with multiple resources a new field
`aws.cloudtrail.multiple_resources` is introduced. If this is set to true, one
does know that multiple resources are associated with this event and it is possible to
look those up in the `event.original`.
But this case is really rare.

So in ~99% of the Cloudtrail events you can easily filter for one resource now. And for the ~1%
of the remaining events you can just filter for the first of those resources but you now know that
there are multiple resources associated.